pyproject.toml
[project]
name = "trading-rl-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "dotenv>=0.9.9",
    "gymnasium>=1.2.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "stable-baselines3[extra]>=2.7.1",
    "ta-lib>=0.6.8",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.2",
    "ruff>=0.14.8",
]

[tool.setuptools.packages.find]
include = ["trading_rl*"]

[tool.pytest.ini_options]
minversion = "9.0"
addopts = ["-ra"]
testpaths = [
    "tests",
]
pythonpath = [
    "."
]
